requirements.txt
requests==2.31.0
arcade==2.6.17
RPi.GPIO==0.7.0
pydub==0.25.1
numpy==1.19.5
sounddevice==0.4.6
pymysql==1.1.0
cn2an==0.5.20
duckduckgo_search==5.0
flask==2.3.2
SpeechRecognition==3.10.0
pyaudio==0.2.11
websocket-client==1.7.0
paho-mqtt==1.6.1
loguru==0.7.2
azure-cognitiveservices-speech==1.42.0
pvporcupine==3.0.5
pvrecorder==1.2.5
adafruit-circuitpython-dht==4.0.8